+++ crates/prelude/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Major
 
+- Migrate `platform::update` to use `Error` instead of `usize`
 - Remove custom error types and use `wasefire-error` instead
 - Rename `native` to `test` and use `native` for native applets